Add files via upload
Write a Java Program that does the following
a. Create a super class called Car. The Car class has the following fields and
 methods. int speed; double regularPrice; String color; double getSalePrice();
b. Create a sub class of Car class and name it as Truck. The Truck class has the 
 following fields and methods. int weight; double getSalePrice(); /If 
 weight>2000,10% discount. Otherwise,20%discount.
c. Create a subclass of Car class and name it as Ford. The Ford class has the following 
 fields and methods int year; int manufacturerDiscount; double getSalePrice();
 //From the sale price computed from Carclass, subtract the manufacturer
 Discount.
d. Create a subclass of Car class and name it as Sedan. The Sedan
 class has the following fields and methods. int length; double getSalePrice(); //If 
 length>20 feet, 5% discount, Otherwise, 10% discount.
e. Create MyOwnAutoShop class which contains the main() method. Perform the 
 following within the main() method.
f. Create an instance of Sedan class and initialize all the fields with appropriate 
 values.
g. Use super(...) method in the constructor for initializing the fields of the 
 superclass.
h. Create an instance of the Ford class and initialize all the fields with appropriate
 values.
i. Use super(...) method in the constructor for initializing the fields of the super 
 Class.
j. Create an instance of Car class and initialize all the fields with appropriate 
 values. Display the sale prices of all instances.